+++ src/structures/vroom/tw_route.cpp
@@ -1156,6 +1156,12 @@ template void TWRoute::replace(const Input& input,
                                const Index last_rank);
 template void
 TWRoute::replace(const Input& input,
+                 const std::vector<Index>::const_iterator first_job,
+                 const std::vector<Index>::const_iterator last_job,
+                 const Index first_rank,
+                 const Index last_rank);
+template void
+TWRoute::replace(const Input& input,
                  const std::vector<Index>::reverse_iterator first_job,
                  const std::vector<Index>::reverse_iterator last_job,
                  const Index first_rank,