+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 - `setup` key for tasks to refine service time modeling (#358)
 - `max_tasks` key limiting route size at vehicle level (#421, #566)
 - Support for custom cost matrices (#415)
-- number of routes in solution summary (#524)
+- Number of routes in solution summary (#524)
+- Implementation for extended SWAP* local search operator (#507)
 - Github Actions CI (#436)
 - Check for libvroom example build in CI (#514)
 
@@ -19,6 +20,7 @@
 ### Removed
 
 - Travis CI builds (#436)
+- Exchange local search operator (#507)
 
 ### Fixed